+++ src/data/iter/reexports.rs
@@ -217,9 +217,9 @@ reexport! { rust: core::iter,
 
 // These are re-exported as methods of the Iter namespace.
 
-// WAIT: https://github.com/rust-lang/rust/pull/135687
+// WAIT:1.86 [re-export FromCoroutine](https://github.com/rust-lang/rust/pull/135687)
 reexport! { rust: core::iter,
-    extra_features: "nightly_coro",
+    extra_flags:(nightly_coro),
     doc: "Creates an iterator where each iteration calls the provided coroutine.",
     @from_coroutine as iter_from_coroutine
 }