+++ src/lib.rs
@@ -30,11 +30,12 @@
 //
 // nightly
 //
+// (In sync with Cargo.toml::[lints.rust.unexpected_cfgs] & build/features.rs::FLAGS_NIGHTLY)
+#![cfg_attr(nightly_coro, feature(coroutines, coroutine_trait, iter_from_coroutine))]
 // (In sync with Cargo.toml::nightly & build/features.rs::NIGHTLY)
 #![cfg_attr(feature = "nightly_allocator", feature(allocator_api))]
 #![cfg_attr(feature = "nightly_autodiff", feature(autodiff))]
 #![cfg_attr(feature = "nightly_bigint", feature(bigint_helper_methods))]
-#![cfg_attr(feature = "nightly_coro", feature(coroutines, coroutine_trait, iter_from_coroutine))]
 #![cfg_attr(feature = "nightly_doc", feature(doc_cfg, doc_notable_trait))]
 #![cfg_attr(all(feature = "nightly_doc", miri), allow(unused_attributes))]
 #![cfg_attr(all(feature = "nightly_doc", not(doc)), allow(unused_attributes))]