+++ src/work/future/coroutine/mod.rs
@@ -5,8 +5,8 @@
 
 mod coro;
 
-#[cfg(feature = "nightly_coro")]
-#[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "nightly_coro")))]
+#[cfg(nightly_coro)]
+#[cfg_attr(feature = "nightly_doc", doc(cfg(nightly_coro)))]
 mod reexports;
 
 crate::items! { // structural access: _mods, _all, _always
@@ -18,8 +18,8 @@ crate::items! { // structural access: _mods, _all, _always
     mod _mods {
         pub use super::coro::*;
 
-        #[cfg(feature = "nightly_coro")]
-        #[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "nightly_coro")))]
+        #[cfg(nightly_coro)]
+        #[cfg_attr(feature = "nightly_doc", doc(cfg(nightly_coro)))]
         pub use super::reexports::*;
     }
     pub(super) mod _all { #[allow(unused)]
@@ -27,7 +27,7 @@ crate::items! { // structural access: _mods, _all, _always
         pub use super::_mods::*;
     }
     pub(super) mod _always { #![allow(unused)]
-        #[cfg(feature = "nightly_coro")]
+        #[cfg(nightly_coro)]
         pub use super::reexports::*;
     }
 }