+++ src/work/future/coroutine/reexports.rs
@@ -6,12 +6,12 @@
 use crate::reexport;
 
 reexport! { rust: core::ops,
-    extra_features: "nightly_coro",
+    extra_flags:(nightly_coro),
     doc: "The trait implemented by builtin coroutine types.",
     Coroutine
 }
 reexport! { rust: core::ops,
-    extra_features: "nightly_coro",
+    extra_flags:(nightly_coro),
     doc: "The result of a coroutine resumption.",
     CoroutineState
 }